+++ packages/server/src/storage/sql/migrations.ts
@@ -1,0 +1,3 @@
+import * as v1 from './migrations/v1.js';
+
+export default { v1 };