+++ packages/store/vitest.config.ts
@@ -8,6 +8,6 @@ export default defineConfig({
 		setupFiles: ['src/__tests__/setup/indexedDB.ts'],
 	},
 	resolve: {
-		conditions: ['development', 'import'],
+		conditions: ['development', 'default'],
 	},
 });